Added function to check link
Added a small function to check link fetched from database if it has 'http://' or 'https://'. Since if the link doesn't have it, then the redirect function will try to redirect to a another route like to `<domain/ip>/<link>`. Example if there is a link `example.com`, the redirect function will try to redirect to `<domain/ip>/example.com` instead of 'example.com'. So, this function will check if it is there or not, and add it if necessary.